op-challenger: Carefully rearrange modules to avoid cycles.
